test/requirements.txt
# Blizzard API tests – run: pip install -r requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
pytest>=7.0.0
